// src/schemas.ts
export const tokenConfigSchema = z.object({
    duration: z.number(),
    secret: z.string().optional()
});

export const emailTemplateSchema = z.object({
    subject: z.string(),
    body: z.string()
});

export const otpConfigSchema = z.object({
    enabled: z.boolean(),
    duration: z.number(),
    length: z.number(),
    emailTemplate: emailTemplateSchema
});

export const mfaConfigSchema = z.object({
    enabled: z.boolean(),
    duration: z.number(),
    rule: z.string()
});

export const passwordAuthConfigSchema = z.object({
    enabled: z.boolean(),
    identityFields: z.array(z.string())
});

export const oAuth2ProviderSchema = z.object({
    pkce: z.boolean().optional(),
    clientId: z.string(),
    name: z.string(),
    clientSecret: z.string(),
    authURL: z.string(),
    tokenURL: z.string(),
    userInfoURL: z.string(),
    displayName: z.string(),
    extra: z.record(z.string(), z.any()).optional()
});

export const oAuth2ConfigSchema = z.object({
    enabled: z.boolean(),
    mappedFields: z.record(z.string(), z.string()),
    providers: z.array(oAuth2ProviderSchema)
});

export const collectionSchema = baseModelSchema.extend({
    name: z.string(),
    fields: z.array(collectionFieldSchema),
    indexes: z.array(z.string()),
    system: z.boolean(),
    listRule: z.string().optional(),
    viewRule: z.string().optional(),
    createRule: z.string().optional(),
    updateRule: z.string().optional(),
    deleteRule: z.string().optional()
});

export const baseCollectionModelSchema = collectionSchema.extend({
    type: z.literal("base")
});

export const viewCollectionModelSchema = collectionSchema.extend({
    type: z.literal("view"),
    viewQuery: z.string()
});

export const authAlertConfigSchema = z.object({
    enabled: z.boolean(),
    emailTemplate: emailTemplateSchema
});

export const authCollectionModelSchema = collectionSchema.extend({
    type: z.literal("auth"),
    authRule: z.string().optional(),
    manageRule: z.string().optional(),
    authAlert: authAlertConfigSchema,
    oauth2: oAuth2ConfigSchema,
    passwordAuth: passwordAuthConfigSchema,
    mfa: mfaConfigSchema,
    otp: otpConfigSchema,
    authToken: tokenConfigSchema,
    passwordResetToken: tokenConfigSchema,
    emailChangeToken: tokenConfigSchema,
    verificationToken: tokenConfigSchema,
    fileToken: tokenConfigSchema,
    verificationTemplate: emailTemplateSchema,
    resetPasswordTemplate: emailTemplateSchema,
    confirmEmailChangeTemplate: emailTemplateSchema
});
// ...
export const collectionModelSchema = z.union([baseCollectionModelSchema, viewCollectionModelSchema, authCollectionModelSchema]);